% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdash.r
\name{Monitor}
\alias{Monitor}
\title{Monitor monitors a machine learning job}
\usage{
Monitor(func, job.name)
}
\arguments{
\item{func}{A function that when invoked, will execute the job you want to monitor.
Note: The provided function must accept an argument called hd.client
which will be passed into the function and exposes various Hyperdash
functionality. For example, if you want to print something, but also have
it available in your Hyperdash logs, you can call hd.client$print("Your log message here")}

\item{job.name}{The name of the job that you want to monitor.}
}
\description{
This function helps you monitor a machine learning job. It will
keep track of when the job starts/ends, as well as any logs that
are emitted using the hd.client.print() function.
}
\examples{
Monitor(function(hd.client) {
  hd.client$print("Begining machine learning...")
  Sys.sleep(2)
  hd.client$print("25\% complete...")
  Sys.sleep(2)
  hd.client$print("50\% complete...")
  Sys.sleep(2)
  hd.client$print("75\% complete...")
  Sys.sleep(2)
  hd.client$print("100\% complete...")
  Sys.sleep(2)
  hd.client$print("Done!")
}, "My test hyperdash job")
}
