% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdash.r
\name{MonitorJob}
\alias{MonitorJob}
\title{MonitorJob monitors a machine learning job}
\usage{
MonitorJob(func, job.name)
}
\arguments{
\item{func}{A function that when invoked, will execute the job you want to monitor.
Note: The provided function must accept an argument called hd_client
which will be passed into the function and exposes various Hyperdash
functionality. For example, if you want to print something, but also have
it available in your Hyperdash logs, you can call hd_client$print("Your log message here")}

\item{job.name}{The name of the job that you want to monitor.}
}
\description{
This function helps you monitor a machine learning job. It will
keep track of when the job starts/ends, as well as any logs that
are emitted using the hd_client.print() function.
}
\examples{
MonitorJob(function(hd_client) {
  hd_client$print("Begining machine learning...")
  Sys.sleep(2)
  hd_client$print("25\% complete...")
  Sys.sleep(2)
  hd_client$print("50\% complete...")
  Sys.sleep(2)
  hd_client$print("75\% complete...")
  Sys.sleep(2)
  hd_client$print("100\% complete...")
  Sys.sleep(2)
  hd_client$print("Done!")
}, "My test hyperdash job")
}
